Update NETStandard and PCL for Android SDK 12.2
