+++ airship.properties
@@ -1,4 +1,4 @@
-crossPlatformVersion = 12.0.0
+crossPlatformVersion = 13.0.0
 iosVersion = 13.1.0
 androidVersion = 12.2.0
