--- src/AirshipBindings.NETStandard/AirshipBindings.NETStandard.Android/packages.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Xamarin.Android.Arch.Core.Common" version="1.0.0" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Arch.Lifecycle.Common" version="1.0.3" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Arch.Lifecycle.Runtime" version="1.0.3" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Support.Annotations" version="27.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Support.Compat" version="27.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Support.Core.UI" version="27.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Support.Core.Utils" version="27.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Support.Fragment" version="27.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Android.Support.Media.Compat" version="27.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Build.Download" version="0.4.9" targetFramework="monoandroid81" />
-</packages>
