+++ src/AirshipBindings.Portable/AirshipBindings.Portable.Abstractions/AirshipBindings.Portable.Abstractions.csproj
@@ -29,16 +29,16 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="IAirship.cs" />
-    <Compile Include="Push\TagEditor.cs" />
-    <Compile Include="Push\TagGroupsEditor.cs" />
+    <Compile Include="Channel\TagEditor.cs" />
+    <Compile Include="Channel\TagGroupsEditor.cs" />
     <Compile Include="Analytics\CustomEvent.cs" />
     <Compile Include="..\..\SharedAssemblyInfo.CrossPlatform.cs">
       <Link>Properties\SharedAssemblyInfo.CrossPlatform.cs</Link>
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Push\" />
+    <Folder Include="Channel\" />
     <Folder Include="Analytics\" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
-</Project>
+</Project>
